requirements.txt
click==8.1.3
colorama==0.4.5
DateTime==4.7
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mysqlclient==2.1.1
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
uuid==1.30
Werkzeug==2.2.2
zope.interface==5.4.0
